Implement Bubble environments initializations methods
