+++ GliaWidgets/Sources/Component/Bubble/BubbleView.Environment.swift
@@ -1,0 +1,30 @@
+import Foundation
+
+extension BubbleView {
+    struct Environment {
+        var data: FoundationBased.Data
+        var uuid: () -> UUID
+        var gcd: GCD
+        var imageViewCache: ImageView.Cache
+    }
+}
+
+extension BubbleView.Environment {
+    static func create(with environment: ViewFactory.Environment) -> Self {
+        .init(
+            data: environment.data,
+            uuid: environment.uuid,
+            gcd: environment.gcd,
+            imageViewCache: environment.imageViewCache
+        )
+    }
+
+    static func create(with environment: EngagementView.Environment) -> Self {
+        .init(
+            data: environment.data,
+            uuid: environment.uuid,
+            gcd: environment.gcd,
+            imageViewCache: environment.imageViewCache
+        )
+    }
+}
