+++ GliaWidgets/Sources/Component/Bubble/BubbleView.swift
@@ -172,12 +172,3 @@ class BubbleView: BaseView {
         gesture.setTranslation(.zero, in: self)
     }
 }
-
-extension BubbleView {
-    struct Environment {
-        var data: FoundationBased.Data
-        var uuid: () -> UUID
-        var gcd: GCD
-        var imageViewCache: ImageView.Cache
-    }
-}
